Drop config for /run/random-seed
The basic premise of /run is that it is a tmpfs that is freshly mounted during
early boot, possibly even before the root file system is available, and is used
to carry state for the current boot. Thus, it hardly could be used to
propagate random generator state from the previous boot.

I found [1] from 2014, which says:
> On most Linux systems, the entropy pool used by /dev/random is
> initialized at boot [2] using /var/run/random-seed. If this process
> were omitted, then two identical servers using the same distribution
> of Linux may have an identical boot sequence and would therefore
> have an identical /dev/random state.
>
> To prevent this from happening, upon shutdown the current value of
> the /dev/random entropy pool is saved off into /var/run/random-seed,
> and then restored upon boot. Making everyone's entropy pool a little
> different, which is a great design.

There is also the following comment there:
> You have that path wrong -- /var/run is a tmpfs and therefore
> deleted on shutdown. I didn't Google it, but my Ubuntu system
> appears to use /var/lib/urandom/random-seed.

[1] https://security.stackexchange.com/questions/69423/add-a-file-as-entropy-source-for-dev-random
[2] http://www.comptechdoc.org/os/linux/startupman/linux_surandom.html (dead link)

In this policy, the initial commit already has the line for
/var/run/random-seed. But it seems to either have been added for some different
system with a different idea about /run, or it was always a misunderstanding.