Set temporary no-stub resolv.conf file from NetworkManager as net_conf_t
Currently NetworkManager writes
/run/NetworkManager/no-stub-resolv.conf by using mkstemp(), which
creates a temporary file name with a random suffix and then renames
it. The policy doesn't handle this, and the file gets wrongly labeled
as NetworkManager_var_run_t.

NetworkManager is being updated to write the file with a fixed ".tmp"
suffix here:

  https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/2298

Adapt the policy accordingly.